Add support for EL Clients to sim tests (#4710)
* Update sim tests structure to add new nodes

* Add new node to work to test sync

* Add sim test for checkpoint sync

* Update the checkpoint sync root

* Update seconds per slot for the simulation tests

* Update the assertion for attestaion count

* Update expectedMinParticipationRate and expectedMinSyncParticipationRate to 90%

* Update genesis delay

* Increase genesisSlotsDelay for ci

* Increase genesisSlotsDelay for ci

* Update test timeout

* Add support for EL

* Update support for multiple nodes

* Increase tcp inactivity timeout for sim tests

* Add sync assertions for range and checkpoint

* Add an estimated timeout for the sim tests

* Add support to pull geth in CI

* Update the CI for geth image

* Update the geth working directory path

* Fix sync configuration

* Fix linter warnings

* Revert "Increase tcp inactivity timeout for sim tests"

This reverts commit 3b10de87c4a82ce038535222446ae626dc970f56.

* Add deterministic TTD for bellatrix

* Fix geth peer connection

* Add new structure for the sim assertions

* Add new sim assertion framework

* Update error logging

* Add range sync and checkpoint sync to new structure

* Update error message logging

* Fix finality assertion check

* Fix lint error

* Rename few variable names

* Fix linter errors

+++ packages/cli/test/utils/simulation/runner/child_process.ts
@@ -6,7 +6,14 @@ import {JobOptions, Job, Runner, RunnerEvent, RunnerType} from "../interfaces.js
 
 type ChildProcessWithJobOptions = {jobOptions: JobOptions; childProcess: ChildProcess};
 
-const stopChildProcess = async (childProcess: ChildProcess, signal?: "SIGTERM"): Promise<void> => {
+const stopChildProcess = async (
+  childProcess: ChildProcess,
+  signal: NodeJS.Signals | number = "SIGTERM"
+): Promise<void> => {
+  if (childProcess.killed || childProcess.exitCode !== null) {
+    return;
+  }
+
   return new Promise((resolve) => {
     childProcess.on("close", resolve);
     childProcess.kill(signal);
@@ -27,18 +34,29 @@ const startChildProcess = async (jobOptions: JobOptions): Promise<ChildProcess> 
       childProcess.stderr?.pipe(stdoutFileStream);
 
       childProcess.on("error", reject);
-      childProcess.on("exit", (code: number) => {
-        clearInterval(intervalId);
-        stdoutFileStream.close();
-        reject(new Error(`process exited with code ${code}`));
-      });
 
-      const intervalId = setInterval(async () => {
-        if (await jobOptions.health()) {
+      if (jobOptions.health !== undefined) {
+        const intervalId = setInterval(async () => {
+          if (jobOptions.health && (await jobOptions.health())) {
+            clearInterval(intervalId);
+            resolve(childProcess);
+          }
+        }, 1000);
+        childProcess.on("exit", (code: number) => {
           clearInterval(intervalId);
-          resolve(childProcess);
-        }
-      }, 1000);
+          stdoutFileStream.close();
+          reject(new Error(`process exited with code ${code}`));
+        });
+      } else {
+        childProcess.on("exit", (code: number) => {
+          stdoutFileStream.close();
+          if (code > 0) {
+            reject(new Error(`process exited with code ${code}`));
+          } else {
+            resolve(childProcess);
+          }
+        });
+      }
     })();
   });
 };
@@ -76,8 +94,8 @@ export class ChildProcessRunner implements Runner {
       console.log(`Stopping "${id}"...`);
       this.emitter.emit("stopping");
       for (const {jobOptions, childProcess} of childProcesses) {
-        if (jobOptions.cleanup) {
-          await jobOptions.cleanup();
+        if (jobOptions.teardown) {
+          await jobOptions.teardown();
         }
         await stopChildProcess(childProcess);
       }
